fix: cow-shed types
# Summary

CoW-shed was using absolute imports in some places which causes the
generated types to not find the the proper imports.

Changed it to relative.

# Testing

In this lib all tests were passing as before.
The issue was when importing it on watch-tower.

I built it with `yalc`, published locally and installed onto
watch-tower.

After that the build was successful.